[DT][NFC] Refactor encoding utilities. (2/n)
The revision moves the below structs and functions to the Codegen
dialect:
  - MaterializeEncodingValueInfo
  - MaterializeEncodingValueFn
  - lowerSetEncodingOpToPackOp
  - lowerUnsetEncodingToUnpackOp

Additionally, it makes few local functions public and moves them to the
Codegen dialect:
  - transposeInPlace
  - getInnerTileSizesOfr

There are no changes in the implementation. The only difference is that
the encoding lowering functions no longer take the type converter in
inputs. Instead, they take ResolveEncodingInfoFn inputs.
